fixed coderabbit reviewed issues
